+++ Gear/Settings.cs
@@ -43,18 +43,14 @@ namespace Gear.Properties
             //
         }
 
-        /// <summary>
-        ///
-        /// </summary>
+        /// <summary></summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
         private void SettingChangingEventHandler(object sender, System.Configuration.SettingChangingEventArgs e) {
             // Agregar código para administrar aquí el evento SettingChangingEvent.
         }
 
-        /// <summary>
-        ///
-        /// </summary>
+        /// <summary></summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
         private void SettingsSavingEventHandler(object sender, System.ComponentModel.CancelEventArgs e) {